+++ src/Graphics/src/Graphics/Platforms/iOS/PlatformImage.cs
@@ -104,6 +104,9 @@ namespace Microsoft.Maui.Graphics.Platform
 
 		private NSData CreateData(ImageFormat format = ImageFormat.Png, float quality = 1)
 		{
+			if (quality < 0 || quality > 1)
+				throw new ArgumentOutOfRangeException(nameof(quality), "quality must be in the range of 0..1");
+
 			NSData data;
 			switch (format)
 			{
