+++ src/Graphics/tests/DeviceTests/Tests/ImageLoadingServiceTests.cs
@@ -52,44 +52,4 @@ public class ImageLoadingServiceTests
 		Assert.Equal(747, image.Width);
 		Assert.Equal(840, image.Height);
 	}
-
-	[Theory]
-	[InlineData(ImageFormat.Png, 1.0f)]
-	[InlineData(ImageFormat.Png, 0.8f)]
-	[InlineData(ImageFormat.Png, 0.4f)]
-	[InlineData(ImageFormat.Jpeg, 1.0f)]
-	[InlineData(ImageFormat.Jpeg, 0.8f)]
-	[InlineData(ImageFormat.Jpeg, 0.4f)]
-	public async Task CanGetBytesFromImage(ImageFormat format, float quality)
-	{
-		var service = new PlatformImageLoadingService();
-
-		using var stream = await FileSystem.OpenAppPackageFileAsync("dotnet_bot.png");
-		using var image = service.FromStream(stream);
-
-		var bytes = image.AsBytes(format, quality);
-
-		Assert.NotNull(bytes);
-		Assert.NotEmpty(bytes);
-	}
-
-	[Theory]
-	[InlineData(ImageFormat.Png, 1.0f)]
-	[InlineData(ImageFormat.Png, 0.8f)]
-	[InlineData(ImageFormat.Png, 0.4f)]
-	[InlineData(ImageFormat.Jpeg, 1.0f)]
-	[InlineData(ImageFormat.Jpeg, 0.8f)]
-	[InlineData(ImageFormat.Jpeg, 0.4f)]
-	public async Task CanGetStreamFromImage(ImageFormat format, float quality)
-	{
-		var service = new PlatformImageLoadingService();
-
-		using var stream = await FileSystem.OpenAppPackageFileAsync("dotnet_bot.png");
-		using var image = service.FromStream(stream);
-
-		var newStream = image.AsStream(format, quality);
-
-		Assert.NotNull(newStream);
-		Assert.True(newStream.Length > 0);
-	}
 }
